moved state reset to Fractal Provider
